move app paths titles to util file
